Move SVGIcon class to its own module
move to `choices.py`
