+++ wagtailio/utils/choices.py
@@ -1,0 +1,42 @@
+from django.db import models
+
+
+class SVGIcon(models.TextChoices):
+    ARROW = "arrow", "Arrow"
+    BI_ARROW = "bi-arrow", "Bi-arrow"
+    BLOG = "blog", "Blog"
+    BRIEFCASE = "briefcase", "briefcase"
+    BOXED_ARROW = "boxed-arrow", "Boxed-arrow"
+    CHEVRON = "chevron", "Chevron"
+    CODE_FILE = "code-file", "Code-file"
+    EDIT = "edit", "Edit"
+    ENVELOPE = "envolope", "Envolope"
+    EYE = "eye", "Eye"
+    FIRE = "fire", "Fire"
+    GITHUB = "github", "Github"
+    GLOBE = "globe", "Globe"
+    HAND_SHAKE = "hand-shake", "Hand-shake"
+    HEART = "heart", "Heart"
+    ID = "id", "ID"
+    INFORMATION = "information", "Information"
+    INSTITUTE = "institute", "Institute"
+    LISTS = "list", "List"
+    LOCATION = "location", "Location"
+    LOCK = "lock", "Lock"
+    MEDIA_LINK = "media-link", "Media-link"
+    MONEY_BAG = "money-bag", "money-bag"
+    MOON = "moon", "Moon"
+    PEOPLE = "people", "People"
+    PING_PONG = "ping-pong", "Ping-pong"
+    PROFESSIONAL = "professional", "Professional"
+    QUOTE = "quote", "Quote"
+    SLACK = "slack", "Slack"
+    SPYGLASS = "spyglass", "Spyglass"
+    STACK = "stack", "Stack"
+    SUNNY_CLOUD = "sunny-cloud", "Sunny cloud"
+    TEAM = "team", "Team"
+    TOAST = "toast", "Toast"
+    T_SHIRT = "t-shirt", "T-shirt"
+    TWITTER = "twitter", "Twitter"
+    WORDPRESS = "wordpress", "Wordpress"
+    YOUTUBE = "youtube", "youtube"
